backend/requirements.txt
flask>=3.0
flask-cors>=4.0
flask-sqlalchemy>=3.0
pydantic>=2.0
google-genai>=1.0
python-dotenv>=1.0
gunicorn>=21.0
pytest>=8.0
